Handle expanding & freezing frame graphs using touch
